+++ frontend/src/js/rewards_scene.ts
@@ -14,9 +14,14 @@ const RFC_5322 = new RegExp(
 export default class RewardsScene extends Phaser.Scene {
   game!: Game;
   music!: Phaser.Sound.BaseSound;
-  explanation!: Phaser.GameObjects.Text;
+  fontSize!: number;
+  heading!: Phaser.GameObjects.Text;
+  stepOne!: Phaser.GameObjects.Text;
   privacy!: Phaser.GameObjects.Text;
   mailBox!: Phaser.GameObjects.Text;
+  stepTwo!: Phaser.GameObjects.Text;
+  formIta!: Phaser.GameObjects.Text;
+  formDeu!: Phaser.GameObjects.Text;
   backBtn!: Phaser.GameObjects.Text;
 
   constructor() {
@@ -26,74 +31,132 @@ export default class RewardsScene extends Phaser.Scene {
   create(data: { music: Phaser.Sound.BaseSound }) {
     this.music = data.music;
 
-    this.createExplanation();
-    this.createMailBox();
-    this.createPrivacyNotice();
-    this.createBackBtn();
+    this.fontSize = Math.min(
+      this.cameras.main.height * 0.06,
+      this.cameras.main.width * 0.03,
+    );
+
+    this.heading = this.createHeading();
+    this.stepOne = this.createStepOne();
+    this.mailBox = this.createMailBox();
+    this.privacy = this.createPrivacy();
+    this.stepTwo = this.createStepTwo();
+    this.formIta = this.createFormIta();
+    this.formDeu = this.createFormDeu();
+    this.backBtn = this.createBackBtn();
+
+    this.refreshMailBox();
+
+    makeButtonHoverable(this.mailBox);
+    makeButtonHoverable(this.privacy);
+    makeButtonHoverable(this.formIta);
+    makeButtonHoverable(this.formDeu);
+    makeButtonHoverable(this.backBtn);
+
     this.bindEvents();
 
     this.setPauseEnabled(false);
   }
 
-  createExplanation() {
-    const text = "We reward top players weekly!\nEnter your email to compete:";
-    const fontSize = Math.min(this.cameras.main.height * 0.125, 24);
-    this.explanation = this.add
+  createHeading() {
+    const text =
+      "Are you a young South Tyrolean?\nYou can play to win prizes in two steps!";
+    return this.add
       .text(0, 0, text, TEXT_STYLES.BASE)
-      .setFontSize(fontSize)
+      .setFontSize(this.fontSize)
       .setOrigin(0.5, 0)
       .setPosition(this.cameras.main.centerX, this.cameras.main.height * 0.05);
   }
 
+  createStepOne() {
+    const text = "1. Enter your email here:";
+    return this.add
+      .text(0, 0, text, TEXT_STYLES.BASE)
+      .setFontSize(this.fontSize)
+      .setOrigin(0.5, 0)
+      .setPosition(
+        this.cameras.main.centerX,
+        this.heading.getBounds().bottom + this.fontSize * 0.6,
+      );
+  }
+
   createMailBox() {
-    this.mailBox = this.add
+    return this.add
       .text(
         this.cameras.main.centerX,
-        this.cameras.main.centerY,
+        this.stepOne.getBounds().bottom,
         "",
         TEXT_STYLES.BUTTON,
       )
-      .setOrigin(0.5, 1)
-      .setPadding(16);
-    makeButtonHoverable(this.mailBox);
-    this.refreshMailBox();
+      .setFontSize(this.fontSize)
+      .setOrigin(0.5, 0);
   }
 
-  createPrivacyNotice() {
-    const text =
-      "By filling the input you agree to our privacy policy.\nYou can read it HERE 🖋️🤓";
-    const fontSize = Math.min(this.cameras.main.height * 0.125, 16);
-    this.privacy = this.add
-      .text(
-        this.cameras.main.centerX,
-        this.mailBox.getBounds().bottom + 16,
-        text,
-        {
-          ...TEXT_STYLES.BUTTON,
-          testString: text,
-          align: "center",
-        },
-      )
-      .setFontSize(fontSize)
+  createPrivacy() {
+    const verb = this.game.device.os.desktop ? "Click" : "Tap";
+    const text = `Filling in the input you accept our privacy policy. ${verb} read it.`;
+    return this.add
+      .text(this.cameras.main.centerX, this.mailBox.getBounds().bottom, text, {
+        ...TEXT_STYLES.BUTTON,
+        testString: text,
+        align: "center",
+      })
+      .setFontSize(this.fontSize * 0.6)
       .setOrigin(0.5, 0);
-    makeButtonHoverable(this.privacy);
+  }
+
+  createStepTwo() {
+    const text = "2. Register on either form:";
+    return this.add
+      .text(0, 0, text, TEXT_STYLES.BASE)
+      .setFontSize(this.fontSize)
+      .setOrigin(0.5, 0)
+      .setPosition(
+        this.cameras.main.centerX,
+        this.privacy.getBounds().bottom + this.fontSize * 0.6,
+      );
+  }
+
+  createFormIta() {
+    const text = "Italiano";
+    return this.add
+      .text(0, 0, text, TEXT_STYLES.BASE)
+      .setFontSize(this.fontSize)
+      .setOrigin(1, 0)
+      .setPosition(
+        this.cameras.main.centerX - this.fontSize * 0.5,
+        this.stepTwo.getBounds().bottom + this.fontSize * 0.25,
+      );
+  }
+
+  createFormDeu() {
+    const text = "Deutsch";
+    return this.add
+      .text(0, 0, text, TEXT_STYLES.BASE)
+      .setFontSize(this.fontSize)
+      .setOrigin(0, 0)
+      .setPosition(
+        this.cameras.main.centerX + this.fontSize * 0.5,
+        this.stepTwo.getBounds().bottom + this.fontSize * 0.25,
+      );
   }
 
   createBackBtn() {
     const text = "Back to menu";
     const fontSize = Math.min(this.cameras.main.height * 0.25, 32);
-    this.backBtn = this.add
+    return this.add
       .text(0, 0, text, TEXT_STYLES.BUTTON)
       .setFontSize(fontSize)
       .setOrigin(0.5, 1)
       .setPosition(this.cameras.main.centerX, this.cameras.main.height * 0.95);
-    makeButtonHoverable(this.backBtn);
   }
 
   bindEvents() {
     this.mailBox.on("pointerup", this.changeEmail.bind(this));
     this.backBtn.on("pointerup", this.backToWelcome.bind(this));
     this.privacy.on("pointerup", this.openPrivacyPolicy.bind(this));
+    this.formIta.on("pointerup", this.openFormIta.bind(this));
+    this.formDeu.on("pointerup", this.openFormDeu.bind(this));
   }
 
   async changeEmail() {
@@ -142,6 +205,18 @@ export default class RewardsScene extends Phaser.Scene {
 
   openPrivacyPolicy() {
     const url = "https://www.eurac.edu/en/static/privacy-policy-website";
+    window.open(url, "_blank");
+  }
+
+  openFormIta() {
+    const url =
+      "https://docs.google.com/forms/d/e/1FAIpQLSd8z7OTUN_cUWJmQGnsq_tjcnFq0g_fJut6rexIBZrFE68NcQ/viewform";
+    window.open(url, "_blank");
+  }
+
+  openFormDeu() {
+    const url =
+      "https://docs.google.com/forms/d/e/1FAIpQLSfIjqC-3REfQwUHRUCuIkhZ0oa1nQMCr7m5UDJwwhBidl1MDg/viewform";
     window.open(url, "_blank");
   }
 }